band3 wave #5: register 17 meta_band TUs + port supporting headers (WIP, may not build)
Scaffolding for the next matching wave. Build may be broken — agents are
actively fixing it; committing the in-progress state on purpose.

Source/config:
- objects.json: add 17 band3/meta_band TUs (AccomplishmentSetlist,
  AccomplishmentTrainer{,Category}Conditional, ContextChecker, NameGenerator,
  OvershellSlotState, SongSortBy{Artist,Plays,Rank,Recent,Song},
  StoreSongSortNode, UIEvent, UIEventMgr, Utl) as NonMatching.
- splits.txt: pin .text ranges for the new clusters (+971 lines).
- defines_common.py: add /I src/band3 and /I src/network include roots
  (mirrors rb3-Wii) so meta_band/* and net headers resolve from any TU.

Ported/extended headers (rb3-Wii/dc3 oracle):
- obj/Msg.h: full MsgSource class (Sink/EventSink, AddSink/RemoveSink, etc.).
- bandobj/BandCharDesc.h: nested Patch class + Category enum for AssetTypes.
- new bandobj/beatmatch/midi/dsp headers (BandScoreboard, BandTrack,
  CrowdMeterIcon, OvershellDir, ScoreDisplay, Track*, DrumMap, DataEvent,
  PitchDetector) plus os/DiscErrorMgr_Wii.h; assorted header touch-ups.

Tooling/docs:
- fingerprint_match.py: callgraph-triangulation oracle generator (+325 lines).
- tools/ninja-locked: flock wrapper for the build.
- docs/decomp/callgraph-triangulation.md, docs/plans/wave5-session-2026-05-28.md.
- gitignore unified_id_callgraph.json (regenerable oracle, ~1 MB).

+++ src/network/Platform/RootObject.h
@@ -3,10 +3,10 @@
 namespace Quazal {
     class RootObject {
     public:
-        static void *operator new(unsigned long);
-        static void *operator new(unsigned long, const char *, unsigned int);
-        static void *operator new[](unsigned long);
-        static void *operator new[](unsigned long, const char *, unsigned int);
+        static void *operator new(size_t);
+        static void *operator new(size_t, const char *, unsigned int);
+        static void *operator new[](size_t);
+        static void *operator new[](size_t, const char *, unsigned int);
         static void operator delete(void *);
         static void operator delete[](void *);
     };
